+++ lib/src/money_data.dart
@@ -4,7 +4,6 @@
  * Written by Brett Sutton <bsutton@onepub.dev>, Jan 2022
  */
 
-
 import 'package:fixed/fixed.dart';
 import 'package:meta/meta.dart' show sealed, immutable;
 import 'currency.dart';